版本时间:2015.10.24

+++ rd/trunk/Client/cms/Assets/script/staticData/EquipLevelData.cs
@@ -4,9 +4,9 @@ using System.Collections;
 public class EquipLevelData
 {
     public int id;
-    public int strength;
-    public int intelligence;
-    public int speed;
-    public int defense;
-    public int health;
+    public float strength;
+    public float intelligence;
+    public float speed;
+    public float defense;
+    public float health;
 }
